+++ beagle-test/tests/fixtures/odin-golden/15-package-name.odin
@@ -2,4 +2,4 @@
 package core
 
 
-chunk_size: i64 : 16;
+chunk_size: int : 16;
